+++ crates/c-api/src/lib.rs
@@ -11,6 +11,7 @@ mod error;
 mod r#extern;
 mod func;
 mod global;
+mod host_ref;
 mod instance;
 mod linker;
 mod memory;
